+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # RELEASE HISTORY
 
+### 6.4.1 / ?, 2021
+* added support for Grouping API:
+```
+func group(queryBuilder: GroupDataQueryBuilder, responseHandler: ((GroupResult) -> Void)!, errorHandler: ((Fault) -> Void)!)
+
+func getGroupObjectCount(queryBuilder: GroupDataQueryBuilder, responseHandler: ((Int) -> Void)!, errorHandler: ((Fault) -> Void)!)
+```
+
 ### 6.4.0 / August 31, 2021
 * added methods to upload file from url:
 ```
@@ -39,7 +47,7 @@ func findByRole(roleName: String, queryBuilder: DataQueryBuilder, responseHandle
 
 func findByRole(roleName: String, loadRoles: Bool, queryBuilder: DataQueryBuilder, responseHandler: (([BackendlessUser]) -> Void)!, errorHandler: ((Fault) -> Void)!)
 ```
-* added the `reloadCurrentUser` property to the UserService to return the updated currentUser data.
+* added the `reloadCurrentUser` property to the UserService to return the updated currentUser data
 
 ### 6.3.3 / June 02, 2021
 * fixed app init with custom domain to accept domens both with protocol or without